+++ src/Transformer/MetadataUtil/AttributesExtractor.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Implementation\Util;
+namespace Rekalogika\Mapper\Transformer\MetadataUtil;
 
-use Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Implementation\Model\Attributes;
+use Rekalogika\Mapper\Transformer\MetadataUtil\Model\Attributes;
 use Rekalogika\Mapper\Util\ClassUtil;
 use Symfony\Component\PropertyInfo\PropertyReadInfo;
 use Symfony\Component\PropertyInfo\PropertyWriteInfo;