+++ src/Transformer/MetadataUtil/PropertyAccessInfoExtractor.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Implementation\Util;
+namespace Rekalogika\Mapper\Transformer\MetadataUtil;
 
 use Symfony\Component\PropertyInfo\PropertyReadInfo;
 use Symfony\Component\PropertyInfo\PropertyReadInfoExtractorInterface;