+++ src/Transformer/MetadataUtil/PropertyMetadataFactoryInterface.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Implementation\Util;
+namespace Rekalogika\Mapper\Transformer\MetadataUtil;
 
-use Rekalogika\Mapper\Transformer\ObjectToObjectMetadata\Implementation\Model\PropertyMetadata;
+use Rekalogika\Mapper\Transformer\MetadataUtil\Model\PropertyMetadata;
 
 /**
  * @internal